Improve the retry logic for wifi Hal starting failure
The retry logic for wifi Hal starting failure is faulty.
In line 1175 we state that
if WifiStatusCode is not available , we should re-try because
the hal might still be stopping.
So we add a counter to +1 and re-try until the counter is > to
START_HAL_RETRY_TIMES which is 3 or until WifiStatusCode is SUCCESS.
This doesn't work. For some reason this is an infinite loop.

So break this infinite loop if the counter hits 3.

Having an infinite loop to re-try to get the WifiStatusCode is
not what we want.

Fixes infinite loop error:
E WifiVendorHal: getWifiLinkLayerStats(l.937) failed {.code = ERROR_NOT_AVAILABLE, .description = }

Change-Id: I52b3fd1933e188cfa7c12bcc56b7e1e239516033