Improve the retry logic for wifi Hal starting failure
The retry logic for wifi Hal starting failure is faulty.
In line 1175 we state that
if WifiStatusCode is not available , we should re-try because
the hal might still be stopping.
So we add a counter to +1 until the WifiStatusCode is SUCCESS.
What about when the hal is not stopping but the WifiStatusCode
reports that it's not available ?

Break this infinite loop if the counter hits 10.
10 times is enough to finally realise that it doesn't need to
try again.

Having an infinite loop to re-try to get the WifiStatusCode is
not what we want.

Change-Id: I52b3fd1933e188cfa7c12bcc56b7e1e239516033